feat: implement multi-tenant settings configuration, cross-tenant data isolation, security guardrails, employee peer collaboration, and super admin control panels

+++ backend/app/schemas/auth.py
@@ -27,7 +27,7 @@ class UserResponse(BaseModel):
     name: str
     email: str
     role: str
-    reward_points: int
+    reward_points: float
     is_active: bool
     created_at: str
